refactor: reorganize functions and split into separate files
