+++ lib/rules/unsafe-to-chain-command/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./unsafeToChainCommand')
